requirements.txt
flask
numpy
opencv-python-headless
pillow
ultralytics
easyocr
paho-mqtt
python-dotenv
torch
torchvision